backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-multipart
python-dotenv
python-magic
pytesseract
easyocr
pillow
python-docx
docxtpl
docx2pdf
passlib[bcrypt]
python-jose[cryptography]
playwright
apscheduler
httpx
